Moving AdministrationContestCategories.tsx, implementing new changes and delete functionality
